Fix bug sometimes the default-tab setting did not work.
The bug is in combination with jquerytools-tab and jquerytools-history,
which caused the history plugin to trigger a tab reload when the default
tab was loaded because the location.hash was changed.

This behavior resulted in two tabs beeing loaded, the default tab and the
first tab of the view with incosistent behavior which of those tabs was
visible at the end.

The fix is here to terminate history events early as long as no actual
hash is in the URL, which is the case when intially loading the tabbed view.
When tabs are klicked the hash is added to the URL, which makes the history
events no longer beeing terminated and enables the default history behavior,
which is that the browser-back and -forward buttons cause the tabs beeing
reloaded correctly.