Fixed initial tab marker (default-tab).
If a tabbedview store his tab on the view itself (see https://github.com/4teamwork/opengever.core/blob/master/opengever/tabbedview/browser/personal_overview.py#L37 for an example), the `initial` class is written to the action itself. Therefore the default tab functionality does not work proper in this case, because the initial class is not reseted after a default-tab change.