Prevent reload loop when requesting non-existing tab.
When a tab was requested by appending `#tabname`, but the tab did not
exist, a reload loop occured.
We no verify whether the tab actually exists.
If it does not exist, the default tab is loaded.

By moving the anchor update from the click event handler to the tab
loading mechanism, we can also "fix" the anchor when it contains no
valid tabname.