Fix tabbedview history
The tabbedview inserts a hash representing the current tab in the URL.
The browser stores this change as an additional history entry beside
the initial URL without any tab selected.
This additional entry where no tab is selected produces a strange behavior
when the user triggers the back button in the browser.
So the fix triggers an additional browser back to skip the additional history entry.
This trigger is only fired when the hash has changed and points to a state
with no tab selected saying no hash in the URL.