Implement p.a.caching etag value adapter.
The new plone.app.caching etag value adapter named ``tabbedview``
provides the current default tab of the user.
This is important for the cache to flush whenever the user selects a new
default tab.

``ftw.tabbedview`` does not enable the etag value.
The integrator needs to do that by adding the ``tabbedview`` etag to the
ruleset configuration.
See the readme for instructions.