Fix bug which caused unsortable columns to be sortable.
When reading from a grid state, the sortable flag was set to "True"
when the grid state did not contain the sortable flag for a column.

This is a wrong assumption since the column may be configured to be
not sortable.

The fix is to use the sortable flag of the column definition with a
fallback to "True" when the sortable flag is not defined on the column
definition.