Fix "instancemethod not iterable" for empty action lists:
Several of the major_actions and minor_actions methods in listing.py used
the and-or trick incorrectly to emulate a ternary expression, like this:

actions = callable(actions) and list(actions()) or actions()

This lead to situations where, if the action list was implemented as a method
(was callable), and the list was empty, the rightmost part of the expression
would be evaluated, because of the falsy value of the empty list.

This lead to TypeError: 'instancemethod' object is not iterable, causing the
entire tab to break.